requirements.txt
csm-ai
streamlit
openai==0.28